requirements.txt
streamlit
pandas
numpy
scikit-learn
sklearn.cluster
sklearn.metrics
sklearn.preprocessing
matplotlib.pyplot